backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==23.2.0
Authlib==1.3.1
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.16.1
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.1.1
Babel==2.15.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.34.139
botocore==1.34.139
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
coverage==7.6.8
cryptography==42.0.8
defusedxml==0.8.0rc2
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
Django==5.0.6
django-appconf==1.0.6
django-auditlog==3.0.0
django-axes==7.0.0
django-backup==1.0.1
django-cors-headers==4.4.0
django-cryptography==1.1
django-debug-toolbar==4.4.6
django-defender==0.9.8
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
django-filter==24.2
django-health-check==3.18.3
django-imagekit==5.0.0
django-modeltranslation==0.18.11
django-oauth-toolkit==2.3.0
django-storages==1.14.4
django-tenants==3.6.1
djangorestframework==3.15.2
dnspython==2.6.1
drf-spectacular==0.27.0
drf-yasg==1.21.7
email_validator==2.2.0
eventlet==0.36.1
factory-boy==3.3.0
Faker==26.0.0
filelock==3.15.4
flake8==7.1.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grep-ast==0.3.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpagentparser==1.9.5
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
jsonformatter==0.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
keyring==24.3.1
ldap3==2.9.1
libgravatar==1.0.4
litellm==1.41.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.3.0
msal==1.29.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.10
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
passlib==1.7.4
pathspec==0.12.1
pilkit==3.0
pillow==10.4.0
playwright==1.45.0
pluggy==1.5.0
portalocker==2.10.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.18.2
PyNaCl==1.5.0
pyotp==2.8.0
pypandoc==1.13
pyparsing==3.1.2
pypng==0.20220715.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-social-auth==0.3.6
python-socketio==5.11.3
python3-openid==3.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
qrcode==7.4.2
redis==5.2.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
scipy==1.13.1
sentry-sdk==2.19.0
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
speaklater3==1.4
SQLAlchemy==2.0.31
sqlparse==0.5.0
sshtunnel==0.4.0
streamlit==1.36.0
structlog==24.4.0
tenacity==8.4.2
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.44.0
wsproto==1.2.0
WTForms==3.1.2
xlsxwriter==3.2.0
yarl==1.9.4
zipp==3.19.2
zstandard==0.22.0

# PDF Generation
reportlab==4.2.5

# Excel Export
openpyxl==3.1.5

# Monitoring and Performance
statsd>=3.3.0
django-statsd-mozilla>=0.4.0
python-statsd>=2.1.0

# Payment Processing
stripe==8.0.0
